Added update check, plugin will now remind you to update if there's a new version found in chat and in console.
Added permission support.

When updating a plugin, it's common for the new version to have new configuration options that don't exist in the old version, hence why I added support to write new config entries to the old config.yml file.

Added the plugins name in its messages it will now show messages from this plugin as [KPManager] in chat, you can change this in config for whatever you want or completely remove it.