Changed License to: ANTI-CAPITALIST SOFTWARE LICENSE (v 1.4)
 The Anti-Capitalist Software License (ACSL) is a software license towards a world beyond capitalism. This license exists to release software that empowers individuals, collectives, worker-owned cooperatives, and nonprofits, while denying usage to those that exploit labor for profit. 

Most existing licenses, including free and open source licenses, consider qualities like source code availability, ease of use, commercialization, and attribution, none of which speak directly to the conditions under which the software is written. Instead, the ACSL considers the organization licensing the software, how they operate in the world, and how the people involved relate to one another.